Refactored code to use provider pattern.
