+++ api/src/test/java/edu/cornell/mannlib/vitro/webapp/service/CaptchaServiceBeanTest.java
@@ -4,8 +4,6 @@ package edu.cornell.mannlib.vitro.webapp.service;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 import java.io.IOException;
@@ -43,78 +41,6 @@ public class CaptchaServiceBeanTest extends AbstractTestClass {
     }
 
     @Test
-    public void validateReCaptcha_InvalidResponse_ReturnsFalse() {
-        // Given
-        props.setProperty("recaptcha.secretKey", "WRONG_SECRET_KEY");
-
-        // When
-        boolean result = CaptchaServiceBean.validateReCaptcha("invalidResponse");
-
-        // Then
-        assertFalse(result);
-    }
-
-    @Test
-    public void generateChallenge_ValidEasyChallengeGenerated() throws IOException {
-        // Given
-        props.setProperty("nanocaptcha.difficulty", "easy");
-
-        // When
-        CaptchaBundle captchaBundle = CaptchaServiceBean.generateChallenge();
-
-        // Then
-        assertNotNull(captchaBundle);
-        assertNotNull(captchaBundle.getB64Image());
-        assertNotNull(captchaBundle.getCode());
-        assertNotNull(captchaBundle.getCaptchaId());
-    }
-
-    @Test
-    public void generateChallenge_ValidEmptyChallengeGenerated() throws IOException {
-        // Given
-        props.setProperty("nanocaptcha.difficulty", "");
-
-        // When
-        CaptchaBundle captchaBundle = CaptchaServiceBean.generateChallenge();
-
-        // Then
-        assertNotNull(captchaBundle);
-        assertNotNull(captchaBundle.getB64Image());
-        assertNotNull(captchaBundle.getCode());
-        assertNotNull(captchaBundle.getCaptchaId());
-    }
-
-    @Test
-    public void generateChallenge_ValidInvalidDifficultyChallengeGenerated() throws IOException {
-        // Given
-        props.setProperty("nanocaptcha.difficulty", "asdasdasd");
-
-        // When
-        CaptchaBundle captchaBundle = CaptchaServiceBean.generateChallenge();
-
-        // Then
-        assertNotNull(captchaBundle);
-        assertNotNull(captchaBundle.getB64Image());
-        assertNotNull(captchaBundle.getCode());
-        assertNotNull(captchaBundle.getCaptchaId());
-    }
-
-    @Test
-    public void generateChallenge_ValidHardChallengeGenerated() throws IOException {
-        // Given
-        props.setProperty("nanocaptcha.difficulty", "hard");
-
-        // When
-        CaptchaBundle captchaBundle = CaptchaServiceBean.generateChallenge();
-
-        // Then
-        assertNotNull(captchaBundle);
-        assertNotNull(captchaBundle.getB64Image());
-        assertNotNull(captchaBundle.getCode());
-        assertNotNull(captchaBundle.getCaptchaId());
-    }
-
-    @Test
     public void getChallenge_MatchingCaptchaIdAndRemoteAddress_ReturnsCaptchaBundle() {
         // Given
         String captchaId = "sampleCaptchaId";
@@ -139,24 +65,20 @@ public class CaptchaServiceBeanTest extends AbstractTestClass {
     }
 
     @Test
-    public void addCaptchaRelatedFieldsToPageContext_RecaptchaImpl() throws IOException {
+    public void addCaptchaRelatedFieldsToPageContext_recaptchaImpl() throws IOException {
         // Given
         props.setProperty("captcha.implementation", "RECAPTCHAv2");
-        props.setProperty("recaptcha.siteKey", "SITE_KEY");
         Map<String, Object> context = new HashMap<>();
 
         // When
         CaptchaServiceBean.addCaptchaRelatedFieldsToPageContext(context);
 
-        // Assert
-        assertNotNull(context.get("siteKey"));
-        assertNull(context.get("challenge"));
-        assertNull(context.get("challengeId"));
+        // Then
         assertEquals("RECAPTCHAV2", context.get("captchaToUse"));
     }
 
     @Test
-    public void addCaptchaRelatedFieldsToPageContext_NanocaptchaImpl() throws IOException {
+    public void addCaptchaRelatedFieldsToPageContext_nanocaptchaImpl() throws IOException {
         // Given
         props.setProperty("captcha.implementation", "NANOCAPTCHA");
         Map<String, Object> context = new HashMap<>();
@@ -164,10 +86,7 @@ public class CaptchaServiceBeanTest extends AbstractTestClass {
         // When
         CaptchaServiceBean.addCaptchaRelatedFieldsToPageContext(context);
 
-        // Assert
-        assertNull(context.get("siteKey"));
-        assertNotNull(context.get("challenge"));
-        assertNotNull(context.get("challengeId"));
+        // Then
         assertEquals("NANOCAPTCHA", context.get("captchaToUse"));
     }
 
@@ -210,20 +129,6 @@ public class CaptchaServiceBeanTest extends AbstractTestClass {
     }
 
     @Test
-    public void validateCaptcha_NanoCaptchaValid() {
-        // Given
-        CaptchaBundle sampleChallenge = new CaptchaBundle("sampleB64Image", "validCode", "challengeId");
-        CaptchaServiceBean.getCaptchaChallenges().put("challengeId", sampleChallenge);
-        props.setProperty("captcha.implementation", "NANOCAPTCHA");
-
-        // Act
-        boolean result = CaptchaServiceBean.validateCaptcha("validCode", "challengeId");
-
-        // Assert
-        assertTrue(result);
-    }
-
-    @Test
     public void validateCaptcha_NoneValid() {
         // Given
         props.setProperty("captcha.enabled", "false");
@@ -231,7 +136,7 @@ public class CaptchaServiceBeanTest extends AbstractTestClass {
         // Act
         boolean result = CaptchaServiceBean.validateCaptcha("anyInput", "anyChallengeId");
 
-        // Assert
+        // Then
         assertTrue(result);
     }
 }
